debugging

+++ src/main/java/com/example/group2backend/database/entity/Comment.java
@@ -9,7 +9,7 @@ import java.time.ZonedDateTime;
 @Data
 public class Comment {
     private Long id; // Comment ID
-    private Integer gameId; // Game ID
+    private Long gameId; // Game ID
     private Long userId; // User ID
     private String content; // Comment content
     private LocalDateTime timestamp; // Comment timestamp
