+++ src/main/java/com/example/group2backend/database/service/CommentService.java
@@ -3,6 +3,7 @@ package com.example.group2backend.database.service;
 import com.example.group2backend.database.entity.Comment;
 import com.example.group2backend.database.mapper.CommentMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
@@ -22,7 +23,7 @@ public class CommentService {
     }
 
     // Get comments by game ID
-    public List<Comment> getCommentsByGameId(Integer gameId) {
+    public List<Comment> getCommentsByGameId(Long gameId) {
         return commentMapper.getCommentsByGameId(gameId);
     }
 }
